chore: theme typing
