+++ app/styles/styled.d.ts
@@ -1,0 +1,6 @@
+import 'styled-components';
+import type { AppTheme } from './theme';
+
+declare module 'styled-components' {
+  export interface DefaultTheme extends AppTheme {}
+}
