+++ app/tsconfig.server.json
@@ -7,6 +7,13 @@
     "isolatedModules": false,
     "noEmit": false
   },
-  "include": ["server.ts", "backend/**/*.ts", "lib/**/*.ts", "data/**/*"],
+  "include": [
+    "server.ts",
+    "backend/**/*.ts",
+    "lib/**/*.ts",
+    "data/**/*",
+    "styles/theme.ts",
+    "styles/styled.d.ts"
+  ],
   "exclude": ["node_modules"]
 }
